Rename DemoFile class to DemoStream
We should make this a stream to be more source agnostic and more
testable. This will also allow us to use live games, if Valve chooses to
stream them in the form of a demo.

I've included a static `DemoStream.FromFile(string path)` method to
replicate the old `new DemoFile(string path);`.
