+++ DemLock.Parser/FrameHandler.cs
